% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ae}
\alias{ae}
\title{ae}
\format{
A dataset with 5 rows and 3 variables:
\describe{
\item{USUBJID}{Unique Subject Identifier}
\item{AETERM}{Reported Term for the Adverse Event}
\item{AESTDTC}{Start Date/Time of Adverse Event}
}
}
\usage{
ae
}
\description{
An example Adverse Events (AE) SDTMv domain.
}
\keyword{data}
