% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_cut.R
\name{apply_cut}
\alias{apply_cut}
\title{Applies the datacut}
\usage{
apply_cut(dsin, dcutvar)
}
\arguments{
\item{dsin}{Name of input dataframe}

\item{dcutvar}{Name of input datacut flagging variable (created by pt_cut and sdtm_cut functions)}
}
\value{
Returns the input dataframe, excluding any rows in which the dcutvar is flagged as "Y". Also removes any
variables with the "DCUT_TEMP" prefix.
}
\description{
Applies the datacut based on the flagging variable created by pt_cut and sdtm_cut functions.
}
\examples{
ae <- data.frame(USUBJID=c("UXYZ123a", "UXYZ123b", "UXYZ123c", "UXYZ123d"),
                 DCUT_TEMP_REMOVE=c("Y", "", "NA", NA))
ae_final <- apply_cut(dsin=ae, dcutvar=DCUT_TEMP_REMOVE)
}
\keyword{derive}
