% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dcut.R
\name{create_dcut}
\alias{create_dcut}
\title{Create DCUT dataset}
\usage{
create_dcut(dataset_ds, filter, cut_date, cut_description)
}
\arguments{
\item{dataset_ds}{Input DS SDTMv dataset}

\item{filter}{Condition to filter patients in DS, should give 1 row per patient}

\item{cut_date}{Datacut date, e.g. "2022-10-22"}
}
\value{
Datacut dataset
}
\description{
Use to create a datacut dataset containing the variables \code{USUBJID}, \code{DCUTDTC}, \code{DCUTDESC}.
}
\examples{
library(admiral)
ds <- tibble::tribble(
 ~USUBJID, ~DSSEQ, ~DSDECOD,
 "subject1", 1, "INFORMED CONSENT",
 "subject1", 2, "RANDOMIZATION",
 "subject1", 3, "WITHDRAWAL BY SUBJECT",
 "subject2", 1, "INFORMED CONSENT",
 "subject3", 1, "INFORMED CONSENT",
 "subject4", 1, "INFORMED CONSENT",
 "subject4", 2, "RANDOMIZATION"
 )

 dcut <- create_dcut(dataset_ds = ds,
                     filter = DSDECOD == "RANDOMIZATION",
                     cut_date = "2022-01-01",
                     cut_description = "Clinical Cutoff Date")
}
\author{
Alana Harris
}
\keyword{derive}
