% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datacutr_dm}
\alias{datacutr_dm}
\title{Demographics SDTMv Dataset}
\format{
A dataset with 5 rows and 3 variables:
\describe{
\item{USUBJID}{Unique Subject Identifier}
\item{DTHFL}{Subject Death Flag}
\item{DTHDTC}{Date/Time of Death}
}
}
\usage{
datacutr_dm
}
\description{
An example Demographics (DM) SDTMv domain.
}
\keyword{data}
