% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datacutr_ds}
\alias{datacutr_ds}
\title{Disposition SDTMv Dataset}
\format{
A dataset with 5 rows and 3 variables:
\describe{
\item{USUBJID}{Unique Subject Identifier}
\item{DSDECOD}{Standardized Disposition Term}
\item{DSSTDTC}{Start Date/Time of Disposition Event}
}
}
\usage{
datacutr_ds
}
\description{
An example Disposition (DS) SDTMv domain.
}
\keyword{data}
