% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datacutr_fa}
\alias{datacutr_fa}
\title{Findings About Events or Interventions SDTMv Dataset}
\format{
A dataset with 5 rows and 4 variables:
\describe{
\item{USUBJID}{Unique Subject Identifier}
\item{FAORRES}{Result or Finding in Original Units}
\item{FADTC}{Date/Time of Collection}
\item{FASTDTC}{Start Date/Time of Observation}
}
}
\usage{
datacutr_fa
}
\description{
An example Findings About Events or Interventions (FA) SDTMv domain.
}
\keyword{data}
