% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_temp_vars.R
\name{drop_temp_vars}
\alias{drop_temp_vars}
\title{Drops Temporary Variables From a Dataset}
\usage{
drop_temp_vars(dsin, drop_dcut_temp = TRUE)
}
\arguments{
\item{dsin}{Name of input dataframe}

\item{drop_dcut_temp}{Whether or not to drop variables beginning with DCUT_TEMP_ (TRUE/FALSE).}
}
\value{
Returns the input dataframe, excluding the temporary variables.
}
\description{
Drops all the temporary variables (variables beginning with TEMP_) from the input dataset.
It also allows the user to specify whether or not to drop the temporary variables needed throughout the
datacut process (variables beginning with DCUT_TEMP_).
}
\details{
The idea of this function is that, when used throughout the datacut process, drop_dcut_temp would
always be set to FALSE until the very end of the datacut process, in which case, drop_dcut_temp would be set
to TRUE.
}
\examples{
test <-  data.frame(USUBJID=c(""), TEMP_XYZ=c(""), DCUT_TEMP_XYZ=c(""))
drop_temp_vars(dsin=test)                           # Drops temp_ and dcut_temp_ variables
drop_temp_vars(dsin=test, drop_dcut_temp=TRUE)    # Drops temp_ and dcut_temp_ variables
drop_temp_vars(dsin=test, drop_dcut_temp=FALSE)   # Drops temp_ variables
}
\keyword{user_utility}
