% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_dcutdtc.R
\name{impute_dcutdtc}
\alias{impute_dcutdtc}
\title{Imputes Partial Dates in DCUTDTC}
\usage{
impute_dcutdtc(dsin, varin, varout)
}
\arguments{
\item{dsin}{Name of input dataframe}

\item{varin}{Name of input datacut date variable}

\item{varout}{Name of imputed output variable}
}
\value{
Returns the input dataframe, with the additional of one extra variable (varout) which is the imputed version of varin.
}
\description{
Imputes partial dates in DCUTDTC variable, created within the datacut process.
}
\examples{
dcut <- data.frame(USUBJID=c("UXYZ123a", "UXYZ123b", "UXYZ123c", "UXYZ123d", "UXYZ123e"),
        DCUTDTC=c("2022-06-23", "2022-06-23T16", "2022-06-23T16:57", "2022-06-23T16:57:30",
                 "2022-06"))
dcut$DCUTDTC <- as.character(dcut$DCUTDTC)
dcut_final <- impute_dcutdtc(dsin=dcut, varin=DCUTDTC, varout=DCUTDT)
}
