% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_sdtm.R
\name{impute_sdtm}
\alias{impute_sdtm}
\title{Imputes Partial Date/Time SDTMv Variables}
\usage{
impute_sdtm(dsin, varin, varout)
}
\arguments{
\item{dsin}{Name of input SDTMv dataframe}

\item{varin}{Name of input SDTMv character date/time variable. Note that this date/time variable must be in ISO 8601
extended format (YYYY-MM-DDThh:mm:ss) and cannot include "fractions of seconds" or "time zones". Any missing components
must be represented by right truncation, rather than the use of additional hyphens. Date/time intervals are not accepted.}

\item{varout}{Name of imputed output variable}
}
\value{
Returns the input SDTMv dataframe, with the addition of one extra variable (varout) in POSIXct datetime format,
which is the imputed version of varin.
}
\description{
Imputes partial date/time SDTMv variables, as required by the datacut process.
}
\examples{
ex <- data.frame(
  USUBJID = c("UXYZ123a", "UXYZ123b", "UXYZ123c", "UXYZ123d"),
  EXSTDTC = c("2022-06-23", "2022-06-23T16", "2022-06-23T16:57", "2022-06-23T16:57:30")
)
temp_ex <- impute_sdtm(dsin = ex, varin = EXSTDTC, varout = DCUT_TEMP_EXSTDTC)
}
\keyword{derive}
