% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lb}
\alias{lb}
\title{lb}
\format{
A dataset with 5 rows and 3 variables:
\describe{
\item{USUBJID}{Unique Subject Identifier}
\item{LBORRES}{Result or Finding in Original Units}
\item{LBDTC}{Date/Time of Specimen Collection}
}
}
\usage{
lb
}
\description{
An example Laboratory Test Results (LB) SDTMv domain.
}
\keyword{data}
