% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_cut.R
\name{pt_cut}
\alias{pt_cut}
\title{Patient Cut}
\usage{
pt_cut(dataset_sdtm, dataset_cut = dcut, cut_var = DCUTDT)
}
\arguments{
\item{dataset_sdtm}{Input SDTMv dataset}

\item{dataset_cut}{Input datacut dataset, default is \code{dcut}}

\item{cut_var}{Datacut date variable found in the \code{dataset_cut} dataset, default is \code{DCUTDT}}
}
\value{
Input dataset plus a flag \code{TEMP_DCUT_REMOVE} to indicate which observations would be dropped when a patient level datacut is applied
}
\description{
Use to apply a datacut a patient cut to an SDTMv dataset
}
\examples{
library(dplyr)
library(admiral.test)

data("admiral_dm")
data("admiral_ae")

dm <- admiral_dm
ae <- admiral_ae

dcut <- dm \%>\%
  filter(RACE == "BLACK OR AFRICAN AMERICAN") \%>\%
  select(USUBJID) \%>\%
  mutate(DCUTDT = "2013-01-01", DCUTDESC = "Clinical Cutoff Date")

ae_out <- pt_cut(ae)
}
\author{
Alana Harris
}
\keyword{derive}
