% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc}
\alias{sc}
\title{sc}
\format{
A dataset with 5 rows and 2 variables:
\describe{
\item{USUBJID}{Unique Subject Identifier}
\item{SCORRES}{Result or Finding in Original Units}
}
}
\usage{
sc
}
\description{
An example Subject Characteristics (SC) SDTMv domain.
}
\keyword{data}
