% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_dm_cut.R
\name{special_dm_cut}
\alias{special_dm_cut}
\title{Special DM Cut}
\usage{
special_dm_cut(
  dataset_dm,
  dataset_cut = dcut,
  cut_var = TEMP_DCUTDT,
  dthcut_var = TEMP_DTHDT
)
}
\arguments{
\item{dataset_dm}{Input DM SDTMv dataset}

\item{dataset_cut}{Input datacut dataset, default is \code{dcut}}

\item{cut_var}{Datacut date variable found in the \code{dataset_cut} dataset, default is \code{TEMP_DCUTDT}}

\item{dthcut_var}{Death date variable (in date format) found in the \code{dataset_dm} dataset, default is \code{TEMP_DMDT}}
}
\value{
A dataset with death information removed if death occurred after data cut
}
\description{
Clears death information within DM if death occurred after datacut date
}
\author{
Tim Barnett
}
\keyword{derive}
