% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts}
\alias{ts}
\title{ts}
\format{
A dataset with 5 rows and 2 variables:
\describe{
\item{USUBJID}{Unique Subject Identifier}
\item{TSVAL}{Parameter Value}
}
}
\usage{
ts
}
\description{
An example Trial Summary (TS) SDTMv domain.
}
\keyword{data}
