% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dcut.R
\name{create_dcut}
\alias{create_dcut}
\title{Create DCUT dataset}
\usage{
create_dcut(dataset_ds, filter, cut_date, cut_description)
}
\arguments{
\item{dataset_ds}{Input DS SDTMv dataset}

\item{filter}{Condition to filter patients in DS, should give 1 row per patient}

\item{cut_date}{Datacut date, e.g. "2022-10-22"}

\item{cut_description}{Datacut date description, e.g. "Clinical Cut Off Date"}
}
\value{
Datacut dataset
}
\description{
Use to create a datacut dataset containing the variables \code{USUBJID}, \code{DCUTDTC}, \code{DCUTDT}, \code{DCUTDESC}.
}
\examples{
ds <- tibble::tribble(
 ~USUBJID, ~DSSEQ, ~DSDECOD, ~DSSTDTC,
 "subject1", 1, "INFORMED CONSENT",      "2020-06-23",
 "subject1", 2, "RANDOMIZATION",         "2020-08-22",
 "subject1", 3, "WITHDRAWAL BY SUBJECT", "2020-05-01",
 "subject2", 1, "INFORMED CONSENT",      "2020-07-13",
 "subject3", 1, "INFORMED CONSENT",      "2020-06-03",
 "subject4", 1, "INFORMED CONSENT",      "2021-01-01",
 "subject4", 2, "RANDOMIZATION",         "2023-01-01"
)

temp_ds <- impute_sdtm(dsin=ds, varin=DSSTDTC, varout=DCUT_TEMP_DSSTDTC)

dcut <- create_dcut(dataset_ds = temp_ds,
                   filter = DSDECOD == "RANDOMIZATION" & DCUTDT>=DCUT_TEMP_DSSTDTC,
                   cut_date = "2022-01-01",
                   cut_description = "Clinical Cutoff Date")
}
\author{
Alana Harris
}
\keyword{derive}
