% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datcutr-package.R
\docType{package}
\name{datacutr-package}
\alias{datacutr}
\alias{datacutr-package}
\title{datacutr: Datacut in R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A toolbox for applying a datacut to SDTMv.
}
\author{
\strong{Maintainer}: Tim Barnett \email{timothy.barnettt@roche.com}

Authors:
\itemize{
  \item Nathan Rees \email{nathan.rees@roche.com}
  \item Alana Harris \email{alana.harris@roche.com}
}

}
\keyword{internal}
