% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_sdtm.R
\name{impute_sdtm}
\alias{impute_sdtm}
\title{Imputes Partial Dates in SDTM Variables}
\usage{
impute_sdtm(dsin, varin, varout)
}
\arguments{
\item{dsin}{Name of input dataframe}

\item{varin}{Name of input SDTM variable}

\item{varout}{Name of imputed output variable}
}
\value{
Returns the input dataframe, with the additional of one extra variable (varout) which
is the imputed version of varin.
}
\description{
Imputes partial dates in SDTM variables as part of the datacut process.
}
\examples{
ex <- data.frame(
  USUBJID = c("UXYZ123a", "UXYZ123b", "UXYZ123c", "UXYZ123d"),
  EXSTDTC = c("2022-06-23", "2022-06-23T16", "2022-06-23T16:57", "2022-06-23T16:57:30")
)
temp_ex <- impute_sdtm(dsin = ex, varin = EXSTDTC, varout = DCUT_TEMP_EXSTDTC)
}
\keyword{derive}
