% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_sdtm.R
\name{impute_sdtm}
\alias{impute_sdtm}
\title{Imputes Partial Dates in SDTM Variables}
\usage{
impute_sdtm(dsin, varin, varout)
}
\arguments{
\item{dsin}{Name of input dataframe}

\item{varin}{Name of input SDTM variable}

\item{varout}{Name of imputed output variable}
}
\value{
Returns the input dataframe, with the additional of one extra variable (varout) which is the imputed version of varin.
}
\description{
Imputes partial dates in SDTM variables as part of the datacut process.
}
