% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cut.R
\name{process_cut}
\alias{process_cut}
\title{Wrapper function to prepare and apply the datacut}
\usage{
process_cut(dataset_cut, cut_var, patient_cut, date_cut, special_dm)
}
\arguments{
\item{dataset_cut}{Input datacut dataset, e.g. dcut}

\item{cut_var}{Datacut date variable, e.g. DCUTDTM}

\item{patient_cut}{A vector of quoted SDTM domain names in which a patient cut will be applied}

\item{date_cut}{A dataframe where column 1 is the SDTM domain names in which a date cut will be applied. Column 2 is the associated SDTM date variable used to carry out the date cut.}

\item{special_dm}{A logical input indicating whether the "special dm cut" should be performed.}
}
\value{
Returns all input SDTM datasets after performing the selected datacut on each domain.
}
\description{
Loops through all input SDTM datasets, applies the selected type of datacut, based on the selected
SDTM date variable, and outputs the resulting datasets
}
\examples{
test <- 1
}
\keyword{derive}
