% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_cut.R
\name{pt_cut}
\alias{pt_cut}
\title{Patient Cut}
\usage{
pt_cut(dataset_sdtm, dataset_cut = dcut, cut_var = DCUTDT, cut_type = "cut")
}
\arguments{
\item{dataset_sdtm}{Input SDTMv dataset}

\item{dataset_cut}{Input datacut dataset, default is \code{dcut}}

\item{cut_var}{Datacut date variable found in the \code{dataset_cut} dataset, default is \code{DCUTDT}}

\item{cut_type}{Set to either "cut" or "drop", "cut" will keep all the patients that are in the datacut dataset
"drop" will keep all the patients that are dropped by applying the patient cut, default is "cut"}
}
\value{
A dataset containing either the records kept or removed after applying the datacut
}
\description{
Apply a patient cut and add a datacut date variable (TEMP_DCUTDT) to an SDTMv dataset
}
\examples{
library(dplyr)
library(lubridate)
library(admiral.test)
library(rlang)
library(purrr)
library(admiral)

data("admiral_dm")
data("admiral_ds")
data("admiral_ex")
data("admiral_ae")
data("admiral_vs")

dm <- admiral_dm
ds <- admiral_ds
ex <- admiral_ex
ae <- admiral_ae
vs <- admiral_vs

dcut <- dm \%>\%
  filter(RACE == "BLACK OR AFRICAN AMERICAN") \%>\%
  select(USUBJID) \%>\%
  mutate(DCUTDT = "2013-01-01", DCUTDESC = "Clinical Cutoff Date")

ae_keep <- pt_cut(ae)
ae_drop <- pt_cut(ae,
                  cut_type = "drop")
dm <- pt_cut(dm)
ds <- pt_cut(ds)
ex <- pt_cut(ex)
ae <- pt_cut(ae)
vs <- pt_cut(vs)


}
\author{
Alana Harris
}
\keyword{assertion}
