% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_cut.R
\name{sdtm_cut}
\alias{sdtm_cut}
\title{xxSTDTC or xxDTC Cut}
\usage{
sdtm_cut(dataset_sdtm, sdtm_date_var, cut_var = TEMP_DCUTDT, cut_type = "cut")
}
\arguments{
\item{dataset_sdtm}{Input SDTMv dataset}

\item{sdtm_date_var}{Input date variable found in the \code{dataset_sdtmv} dataset}

\item{cut_var}{Datacut date variable, default is \code{TEMP_DCUTDT}}

\item{cut_type}{Set to either "cut" or "drop", "cut" will keep all the records before the datacut in
the output dataset, "drop" will keep all the records which occur after and would be dropped, default is
"cut"}
}
\value{
A dataset containing either the records kept or removed after applying the datacut
}
\description{
Apply a datacut to either an xxSTDTC or xxDTC SDTM date variable
}
\examples{
library(dplyr)
library(admiral.test)
library(admiral)

data("admiral_dm")
data("admiral_ae")

dm <- admiral_dm
ae <- admiral_ae

dcut <- dm \%>\%
  filter(RACE == "BLACK OR AFRICAN AMERICAN") \%>\%
  select(USUBJID) \%>\%
  mutate(DCUTDT = "2013-01-01", DCUTDESC = "Clinical Cutoff Date")

ae <- ae \%>\%
  pt_cut(.) \%>\%
  mutate(DCUT_TEMP_AESTDT <- AESTDTC) \%>\%
  sdtm_cut(dataset_sdtm = .,
           sdtm_date_var = AESTDTC)
}
\author{
Alana Harris
}
\keyword{derive}
