% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_dm_cut.R
\name{special_dm_cut}
\alias{special_dm_cut}
\title{Special DM Cut}
\usage{
special_dm_cut(
  dataset_dm,
  dataset_cut = dcut,
  cut_var = DCUTDT,
  dthcut_var = DTHDTC
)
}
\arguments{
\item{dataset_dm}{Input DM SDTMv dataset}

\item{dataset_cut}{Input datacut dataset, default is \code{dcut}}

\item{cut_var}{Datacut date variable found in the \code{dataset_cut} dataset, default is \code{TEMP_DCUTDT}}

\item{dthcut_var}{Death date variable (in date format) found in the \code{dataset_dm} dataset, default is \code{TEMP_DMDT}}
}
\value{
A dataset with death information removed if death occurred after data cut
}
\description{
Clears death information within DM if death occurred after datacut date
}
\examples{


library(dplyr)
library(lubridate)

dcut <- tibble::tribble(
~USUBJID, ~DCUTDTC,
"01-701-1015", "2014-10-20T23:59:59",
"01-701-1023", "2014-10-20T23:59:59",
) \%>\%
  mutate(DCUTDT = ymd_hms(DCUTDTC))

dm <- tibble::tribble(
~USUBJID, ~DTHDTC, ~DTHFL,
"01-701-1015", "2014-10-20", "Y",
"01-701-1023", "2014-10-21", "Y",
)

  special_dm_cut(dataset_dm=dm,
                      dataset_cut=dcut,
                      cut_var=DCUTDT,
                      dthcut_var=DTHDTC)
}
\author{
Tim Barnett
}
\keyword{derive}
