% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_dm_cut.R
\name{special_dm_cut}
\alias{special_dm_cut}
\title{Special DM Cut}
\usage{
special_dm_cut(
  dataset_dm,
  dataset_cut = dcut,
  cut_var = TEMP_DCUTDT,
  dthcut_var = TEMP_DTHDT
)
}
\arguments{
\item{dataset_dm}{Input DM SDTMv dataset}

\item{dataset_cut}{Input datacut dataset, default is \code{dcut}}

\item{cut_var}{Datacut date variable found in the \code{dataset_cut} dataset, default is \code{TEMP_DCUTDT}}

\item{dthcut_var}{Death date variable (in date format) found in the \code{dataset_dm} dataset, default is \code{TEMP_DMDT}}
}
\value{
A dataset with death information removed if death occurred after data cut
}
\description{
Clears death information within DM if death occurred after datacut date
}
\examples{


library(dplyr)
library(stringr)
library(lubridate)
subj <- c("01-701-1015","01-701-1023","01-701-1028","01-701-1033","01-701-1047","01-701-1057","01-701-1097","01-701-1111","01-701-1115","01-701-1118")
cutdt <- c("2014-10-20T23:59:59")
dthdtclist <- c("","2014-10-20","2014-10-21","2013","2014-10","2014-11","2014-09","","","2014-11-20")
dcut <- data.frame(USUBJID=subj,DCUTDTC=cutdt) \%>\%
  mutate(DCUTDT = ymd_hms(DCUTDTC)) \%>\%
  as_tibble()
dm <- data.frame(USUBJID=subj,DTHDTC=dthdtclist) \%>\%
  mutate(DTHFL=case_when(
    DTHDTC!="" ~ "Y",
    TRUE ~ ""
  )) \%>\%
  as_tibble() \%>\%
  mutate(DTHDTC=as.character(DTHDTC))

dm_ptcut <- pt_cut(dm,dcut) \%>\%
  mutate(DCUTDT=TEMP_DCUTDT)
  dm_temp <- impute_sdtm(dm_ptcut,DTHDTC,DCUT_TEMP_DTHDT)

chk <- special_dm_cut(dataset_dm=dm_temp,
                      dataset_cut=dcut,
                      cut_var=DCUTDT,
                      dthcut_var=DCUT_TEMP_DTHDT)
}
\author{
Tim Barnett
}
\keyword{derive}
